requirements.txt
asgiref==3.4.1
cachetools==4.2.4
cffi==1.15.0
cryptography==36.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
google-auth==2.3.3
importlib-metadata==4.8.2
Markdown==3.3.6
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.5.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
zipp==3.6.0
